rename: signinPage -> LoginPage
